PROYECTO 1/requirements.txt
cffi==1.14.2
pycparser==2.20
pycryptodome==3.9.8
python-dotenv==0.14.0
six==1.15.0
